test: consolidate and extend workflow tests
Merge the workflow tests into a single file. Updated prompts to place workflow commands at various positions with trailing whitespace and parameters, and added multi-command scenarios.
